backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-google-genai
langchain-community
langgraph
pinecone-client
langchain-pinecone
pypdf
tiktoken
python-multipart
websockets
requests
beautifulsoup4
textblob
duckduckgo-search
docx2txt
SpeechRecognition
pyaudio

